Gyroscope missing

+++ core/src/Logic/Body/AmmoBody.java
@@ -12,14 +12,14 @@ import Logic.View.GameStageView;
  */
 
 public class AmmoBody extends EntityBody {
-    public AmmoBody(int x, int y, World world, Entity e) {
+    public AmmoBody(float x, float y, World world, Entity e) {
         super(x,y,world,e);
     }
 
 
 
     @Override
-    public void setBody(int x, int y , World world,Entity e) {
+    public void setBody(float x, float y , World world,Entity e) {
         BodyDef bodyDef = new BodyDef();
         bodyDef.type = BodyDef.BodyType.DynamicBody;
         bodyDef.linearVelocity.set(0f,0f);
