+++ core/src/Logic/Body/EntityBody.java
@@ -11,8 +11,8 @@ import Logic.Model.Entity;
 
 public abstract class EntityBody{
     protected Body body;
-    protected int originX, originY;
-    public EntityBody(int x, int y, World world, Entity e){
+    protected float originX, originY;
+    public EntityBody(float x, float y, World world, Entity e){
         setBody(x, y, world, e);
         this.originX = x;
         this.originY = y;
@@ -20,7 +20,7 @@ public abstract class EntityBody{
 
     public Body getBody(){return body;}
 
-    public abstract void setBody(int x, int y, World world, Entity e);
+    public abstract void setBody(float x, float y, World world, Entity e);
 
     public  abstract void update(Entity e);
 
@@ -28,19 +28,19 @@ public abstract class EntityBody{
         this.body = body;
     }
 
-    public int getOriginX() {
+    public float getOriginX() {
         return originX;
     }
 
-    public void setOriginX(int originX) {
+    public void setOriginX(float originX) {
         this.originX = originX;
     }
 
-    public int getOriginY() {
+    public float getOriginY() {
         return originY;
     }
 
-    public void setOriginY(int originY) {
+    public void setOriginY(float originY) {
         this.originY = originY;
     }
